Revert "Import pip2 version in docs"
This reverts commit 61358bb1b96fdb2c8a7f0e2884a26c24fe4979d2.  The
original goal was to avoid duplicating the version by storing it in
pip2/__init__.py and importing it from here.  This caused problems while
trying to put the documentation on Read the Docs, however.  Importing
pip2 ends up trying to import distutils2, which isn't installed when it
tries to build the docs.  For a few different reasons, getting
distutils2 installed in order to build the documentation isn't trivial
(mainly because it isn't stable/released yet).

So, for now, we'll just duplicate the version in a few places.