Test mixed results for install command
Add a test (and update the README) for the case where you are installing
two packages, one the ends of succeeding and one that ends up failing.
The current expected behavior is to allow some packages to be installed
even if there other packages fail to install. In the future, it will
probably be all or nothing. See issue #51 for more details