Tox: Build docs only under Python 2.7
Added an 'rtfd' testenv to tox.ini to verify the documentation is built
properly with Python 2.7 and to gain confidence that it will work on
readthedocs.org.

Removed the documentation build step on other version of Python to speed
up the virtualenv creation (getting/install Sphinx was slow) while
running tox.